requirements.txt
python-chess==1.999
matplotlib==3.10.0
cairosvg==2.7.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
